Add columns variablesFound and variablesNotFound.
These columns will hold the names of the variables of the commands
that were found and not found. This makes it easier to troubleshoot
a command directly from Geneos Active Console without having to SSH
to the netprobe server to find out what error message was displayed.

As such, commands will no longer panic the whole xtender run when
encountering unknown environment variables.