Add decryption of encrypted environment variables.
Geneos has support for encrypted secrets using AES256. This commit adds
support for decrypting these secrets when found in variables using a
provided keyfile.

The option `-k, --key-file` is added to allow the user to provide a
keyfile location. When used, this option will have the highest priority.
If this option is not used, xtender will look for a keyfile at
`/opt/xtender/secret.key`.

If no keyfile is provided xtender will throw an error if it encounters
an encrypted variable.

Note that any keyfile needs to be readable by the user running the
xtender binary.

In the dataview, the secret variable will be substituted with '***'.
However, the output columns are not scanned for the decrypted secret, so
if you echo it out, or if the plugin you run prints it, xtender will
not obfuscate it.